Move genshi._six.add_metaclass to genshi.compat.
