+++ genshi/compat.py
@@ -109,6 +109,27 @@ if IS_PYTHON2:
 else:
     exec_ = getattr(builtins, "exec")
 
+
+# Class decorator for adding a metaclass (from six 1.17.0)
+
+def add_metaclass(metaclass):
+    """Class decorator for creating a class with a metaclass."""
+    def wrapper(cls):
+        orig_vars = cls.__dict__.copy()
+        slots = orig_vars.get('__slots__')
+        if slots is not None:
+            if isinstance(slots, str):
+                slots = [slots]
+            for slots_var in slots:
+                orig_vars.pop(slots_var)
+        orig_vars.pop('__dict__', None)
+        orig_vars.pop('__weakref__', None)
+        if hasattr(cls, '__qualname__'):
+            orig_vars['__qualname__'] = cls.__qualname__
+        return metaclass(cls.__name__, cls.__bases__, orig_vars)
+    return wrapper
+
+
 # We do some scary stuff with CodeType() in template/eval.py
 
 if IS_PYTHON2:
